% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fTA.R
\name{fTA}
\alias{fTA}
\title{Defective Density: P(T_j, A_j) for a Single Subject}
\usage{
fTA(DT, A, S, R, Q, alpha, b, c, tau, beta)
}
\arguments{
\item{DT}{A vector of response times.}

\item{A}{A vector of actions.}

\item{S}{A vector of states.}

\item{R}{A vector of rewards.}

\item{Q}{Initial Q-values matrix.}

\item{alpha}{DDM parameter.}

\item{b}{DDM parameter.}

\item{c}{DDM parameter.}

\item{tau}{DDM parameter.}

\item{beta}{Learning rate in RL.}
}
\value{
A vector of probabilities for each trial.
}
\description{
This function computes the defective density for a single subject.
}
