% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fTA_mat.R
\name{fTA_mat}
\alias{fTA_mat}
\title{Defective Density: P(T_j, A_j) for All Subjects}
\usage{
fTA_mat(T_mat, A_mat, S_mat, R_mat, Q0, alpha, b, c, tau, beta, cores)
}
\arguments{
\item{T_mat}{A matrix of response times.}

\item{A_mat}{A matrix of actions.}

\item{S_mat}{A matrix of states.}

\item{R_mat}{A matrix of rewards.}

\item{Q0}{Initial Q-values matrix.}

\item{alpha}{Parameter for DDM.}

\item{b}{Parameter for DDM.}

\item{c}{Parameter for DDM.}

\item{tau}{Parameter for DDM.}

\item{beta}{Learning rate in RL.}

\item{cores}{Number of cores for parallel processing.}
}
\value{
A matrix of probabilities for all subjects.
}
\description{
This function computes defective densities for all subjects in parallel.
}
