% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_trans.R
\name{inv_trans}
\alias{inv_trans}
\title{Inverse Transformation of Parameters}
\usage{
inv_trans(x, minRT)
}
\arguments{
\item{x}{A vector of transformed parameters.}

\item{minRT}{Minimum response time.}
}
\value{
A vector of inverse-transformed parameters.
}
\description{
This function applies the inverse transformation to the model parameters.
}
