% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_RLDDM.R
\name{obj_RLDDM}
\alias{obj_RLDDM}
\title{Objective Function for RL-DDM Model}
\usage{
obj_RLDDM(pars, T_mat, A_mat, S_mat, R_mat, Q, cores)
}
\arguments{
\item{pars}{A vector of model parameters.}

\item{T_mat}{A matrix of response times.}

\item{A_mat}{A matrix of actions.}

\item{S_mat}{A matrix of states.}

\item{R_mat}{A matrix of rewards.}

\item{Q}{A matrix representing Q-values for the RL model.}

\item{cores}{Number of cores for parallel processing.}
}
\value{
The negative log-likelihood.
}
\description{
This function computes the negative log-likelihood for the RL-DDM model given parameters.
}
