% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_DDM.R
\name{RL_DDM}
\alias{RL_DDM}
\title{RL-DDM Algorithm}
\usage{
RL_DDM(
  T_mat,
  A_mat,
  S_mat,
  R_mat,
  q0,
  alpha0,
  b0,
  c0,
  tau0,
  beta0,
  tol = 1e-04,
  cores = 4
)
}
\arguments{
\item{T_mat}{A matrix of response times.}

\item{A_mat}{A matrix of actions.}

\item{S_mat}{A matrix of states.}

\item{R_mat}{A matrix of rewards.}

\item{q0}{Initial Q-value.}

\item{alpha0}{Initial alpha value.}

\item{b0}{Initial b value.}

\item{c0}{Initial c value.}

\item{tau0}{Initial tau value.}

\item{beta0}{Initial beta value.}

\item{tol}{Tolerance for convergence (default is 1e-4).}

\item{cores}{Number of cores for parallel processing (default is 4).}
}
\value{
A list of optimized parameters (alpha, b, c, tau, beta).
}
\description{
This function runs the RL-DDM algorithm.
}
