% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\title{Transformation of Parameters}
\usage{
trans(x, minRT)
}
\arguments{
\item{x}{A vector of model parameters.}

\item{minRT}{Minimum response time.}
}
\value{
A transformed parameter vector.
}
\description{
This function applies a transformation to the model parameters to ensure valid ranges.
}
