% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Q.R
\name{update_Q}
\alias{update_Q}
\title{This function updates the Q-values (action-state value matrix) based on
the observed reward. The update follows the standard Q-learning rule, where
the Q-value for the selected action in the current state is adjusted
according to the received reward.}
\usage{
update_Q(a, s, R, Q, beta)
}
\arguments{
\item{a}{The action taken (0 or 1).}

\item{s}{The state (0 or 1).}

\item{R}{The observed reward for the action taken in the given state.}

\item{Q}{A 2x2 matrix of Q-values, representing the action-state values.}

\item{beta}{The learning rate.}
}
\value{
The updated Q matrix.
}
\description{
This function updates the Q-values (action-state value matrix) based on
the observed reward. The update follows the standard Q-learning rule, where
the Q-value for the selected action in the current state is adjusted
according to the received reward.
}
