% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_RLDDM.R
\name{update_RLDDM}
\alias{update_RLDDM}
\title{Update for RL-DDM}
\usage{
update_RLDDM(pars, T_mat, A_mat, S_mat, R_mat, Q, cores)
}
\arguments{
\item{pars}{Initial parameter values.}

\item{T_mat}{A matrix of response times.}

\item{A_mat}{A matrix of actions.}

\item{S_mat}{A matrix of states.}

\item{R_mat}{A matrix of rewards.}

\item{Q}{A matrix representing Q-values for the RL model.}

\item{cores}{Number of cores for parallel processing.}
}
\value{
Optimized parameter values.
}
\description{
This function optimizes the parameters of the RL-DDM model by minimizing the negative log-likelihood.
}
